Improvements

+++ app/include/Camera.h
@@ -19,10 +19,10 @@ private:
   float m_scale;
   float m_offsetX;
   float m_offsetY;
-  int m_worldWidth;
-  int m_worldHeight;
-  int m_windowWidth;
-  int m_windowHeight;
+    float m_worldWidth;
+    float m_worldHeight;
+    float m_windowWidth;
+    float m_windowHeight;
 };
 
 #endif // PACMAN_CAMERA_H
